Move compilation tests to a shorted package
